# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Financial management API for cross-border trader
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: tripId
          in: query
          schema:
            type: ["integer", "null"]
        - name: currency
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    patch:
      operationId: updateTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CLIENTS ────────────────────────────────────────────────────
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    patch:
      operationId: updateTrip
      tags: [trips]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{id}/pnl:
    get:
      operationId: getTripPnl
      tags: [trips]
      summary: Get trip profit & loss per currency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip P&L
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripPnl"

  # ─── STUDIOS ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, date, type, amount, currency, status, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        type:
          type: string
          description: "income | expense | payment | receipt"
        amount:
          type: number
        currency:
          type: string
          description: "AED | USD | SYP"
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        tripId:
          type: ["integer", "null"]
        tripName:
          type: ["string", "null"]
        studioId:
          type: ["integer", "null"]
        studioName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
          description: "pending | settled"
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [date, type, amount, currency, status]
      properties:
        date:
          type: string
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        clientId:
          type: ["integer", "null"]
        tripId:
          type: ["integer", "null"]
        studioId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
    TransactionUpdate:
      type: object
      properties:
        date:
          type: string
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        clientId:
          type: ["integer", "null"]
        tripId:
          type: ["integer", "null"]
        studioId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string

    # ─── CLIENT ───────────────────────────────────────────────────
...
    Trip:
      type: object
      required: [id, name, isShared, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        isShared:
          type: boolean
        status:
          type: string
          description: "active | closed"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    TripInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        isShared:
          type: boolean
        status:
          type: string
        notes:
          type: ["string", "null"]
    TripUpdate:
      type: object
      properties:
        name:
          type: string
        isShared:
          type: boolean
        status:
          type: string
        notes:
          type: ["string", "null"]
    TripCurrencyPnl:
      type: object
      required: [currency, revenue, expenses, netProfit, myShare]
      properties:
        currency:
          type: string
        revenue:
          type: number
        expenses:
          type: number
        netProfit:
          type: number
        myShare:
          type: number
    TripPnl:
      type: object
      required: [trip, breakdown]
      properties:
        trip:
          $ref: "#/components/schemas/Trip"
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/TripCurrencyPnl"

    # ─── STUDIO ───────────────────────────────────────────────────
...