# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Financial management API for cross-border trader
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionUpdate {
  date?: string;
  type?: string;
  amount?: number;
  currency?: string;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  tripId?: number | null;
  /** @nullable */
  studioId?: number | null;
  /** @nullable */
  description?: string | null;
  status?: string;
}