# lib/api-zod/src/generated/types/tripCurrencyPnl.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Financial management API for cross-border trader
 * OpenAPI spec version: 0.1.0
 */

export interface TripCurrencyPnl {
  currency: string;
  revenue: number;
  expenses: number;
  netProfit: number;
  myShare: number;
}